Library working.
Picking bits one by one - real usage will show if it is fast enough.
A convenient function fixSign can be used to make micro regard a parameters
sign correctly.

E.g. Latitude consisting of just 27bits. If it is negative and being
handled as a 32bit (long) all bits 27..31 must be set to make it a
negative number