Examples/linux: Added simple demo to be used on linux host
Right now I regret I've ever made this module as an
Arduino module. Really no need for that. Would be so
easy to make it a plain generic C++ module to be used
at all platforms.
Unfortunately I have some time constraints in my life
so made this quick and dirty to make some easy debugging.